+++ weblate/trans/models/source.py
@@ -66,3 +66,10 @@ class Source(models.Model):
             old = Source.objects.get(pk=self.pk)
             self.priority_modified = (old.priority != self.priority)
         super(Source, self).save(force_insert, **kwargs)
+
+    @models.permalink
+    def get_absolute_url(self):
+        return ('review_source', (), {
+            'project': self.subproject.project.slug,
+            'subproject': self.subproject.slug,
+        })